optimization

+++ 15-16_weeks/real-world-gas-optimization/01_synthetix_staking_reward_forge/src/interfaces/IStakingRewards.sol
@@ -12,9 +12,9 @@ interface IStakingRewards {
 
     function getRewardForDuration() external view returns (uint256);
 
-    function lastTimeRewardApplicable() external view returns (uint256);
+    function lastTimeRewardApplicable() external view returns (uint64);
 
-    function rewardPerToken() external view returns (uint256);
+    function rewardPerToken() external view returns (uint64);
 
     function rewardsDistribution() external view returns (address);
